Fetching posts on Home page

+++ admin/delete-category.php
@@ -9,12 +9,21 @@ if (isset($_GET['id'])) {
     // FOR LATER
     // update category_id of posts that belong to this category to id of uncategorized category
 
+    $update_query = "UPDATE posts SET category_id=5 WHERE category_id=$id";
+    $update_result = mysqli_query($connection, $update_query);
 
-    // delete category
-    $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
-    $result = mysqli_query($connection,$query);
-    $_SESSION['delete-category-success'] = "Category deleted successfully";
+    if (!mysqli_errno($connection)) {
+        // delete category
+        $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
+        $result = mysqli_query($connection,$query);
+        $_SESSION['delete-category-success'] = "Category deleted successfully";
 
+    }
+
+
+
+
+    
 
 }
